added customer character and updated KitchenMixer
Created the Customer Scene Base Node2D with children AnimatedSprite, and Area2D. Area2D has child CollisionShape2D in a capsule shape.  Set 8 animations 4 walking and 4 idle to be used later after movement is added.
Updated SellPoint Collision RigidBody2D collision shape to be a bit larger and set the whole scenes Z axis to 2 so that it will appear a little bit over the character like they are standing behind it.
Updated PrepStation KitchenMixer extensively. Added Timer named KitchenMixerTimer, Particles2D and several new functions to trigger the animation to play and particles to show on the start of the Timer and stop when the Timer ends. Had issues with accessing the array that denotes which workspace is open to receive a new item, so I broke down the one function i had before into 4 and used some array native methods to check and trigger other functions.